Merge Greg's newest changes to fix master

+++ kidsapp-api/routes/student.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var Student = require('../models').Student;
+var Class = require('../models').Class;
 
 module.exports.getAll = function (req, res) {
   Student.findAll({
@@ -9,7 +10,7 @@ module.exports.getAll = function (req, res) {
     }
   }).then(function (students) {
     if (students.length === 0) {
-      res.status(404).json({message: "no students found"});
+      res.status(404).json({message: "No students found"});
     }
     else {
       var responseObj = [];
@@ -34,7 +35,7 @@ module.exports.getById = function (req, res) {
     }
   }).then(function (student) {
     if (!student) {
-      res.status(404).json({"message": "student not found"});
+      res.status(404).json({"message": "Student not found"});
     }
     else {
       res.json({
@@ -52,23 +53,34 @@ module.exports.post = function (req, res) {
     if (!req.body.username || !req.body.firstName || !req.body.lastName || !req.body.password) {
       res.status(400).json({message: "Invalid student request format."});
     }
-
-    var student = Student.build({
-      username : req.body.username,
-      firstName : req.body.firstName,
-      lastName : req.body.lastName,
-      password : req.body.password,
-      ClassId : req.params.classId
-    });
-
-    student.save()
-      .then(function () {
-        res.json(
-          {message: "Inserted student successfully"});
-      })
-      .error(function (err) {
-        throw err;
+    else{
+      Class.findOne({
+        where: {
+          id : req.params.classId
+        }
+      }).then(function (foundClass) {
+        if (!foundClass) {
+          res.status(404).json({message: "Class not found"});
+        }
+        else {
+          var student = Student.build({
+            username : req.body.username,
+            firstName : req.body.firstName,
+            lastName : req.body.lastName,
+            password : req.body.password,
+            ClassId : req.params.classId
+          });
+          student.save()
+            .then(function () {
+              res.json(
+                {message: "Inserted student successfully"});
+            })
+            .error(function (err) {
+              throw err;
+            });
+        }
       });
+    }
   }
   catch (err) {
     console.log(err);
@@ -81,30 +93,42 @@ module.exports.put = function (req, res) {
     if (!req.body.username || !req.body.firstName || !req.body.lastName || !req.body.password) {
       res.status(400).json({message: "Invalid student request format."});
     }
-
-    Student.update(
-      {
-        username: req.body.username,
-        lastName: req.body.lastName,
-        firstName: req.body.firstName,
-        password: req.body.password
-      },
-      {
+    else{
+      Class.findOne({
         where: {
-          username : req.params.studentId,
-          ClassId : req.params.classId
+          id : req.params.classId
         }
-      })
-      .then(function (updated) {
-        if (updated > 0) {
-          res.json({message: "Student updated"});
+      }).then(function (foundClass) {
+        if (!foundClass) {
+          res.status(404).json({message: "Class not found"});
         }
         else {
-          res.status(404).json({message: "Student not found"});
+          Student.update(
+            {
+              username: req.body.username,
+              lastName: req.body.lastName,
+              firstName: req.body.firstName,
+              password: req.body.password
+            },
+            {
+              where: {
+                username : req.params.studentId,
+                ClassId : req.params.classId
+              }
+            })
+            .then(function (updated) {
+              if (updated > 0) {
+                res.json({message: "Student updated"});
+              }
+              else {
+                res.status(404).json({message: "Student not found"});
+              }
+            }).error(function (err) {
+            throw err;
+          });
         }
-      }).error(function (err) {
-      throw err;
-    });
+      });
+    }
   }
   catch (e) {
     console.error(e);
@@ -114,20 +138,31 @@ module.exports.put = function (req, res) {
 
 module.exports.delete = function (req, res) {
   try {
-    Student.destroy({
-        where: {
-          username : req.params.studentId,
-          ClassId : req.params.classId
-        }
-      })
-      .then(function (result) {
-        if (!result) {
-          res.status(404).json({message: "Student not found"});
-        }
-        else {
-          res.json({message: "Student deleted"});
-        }
-      });
+    Class.findOne({
+      where: {
+        id : req.params.classId
+      }
+    }).then(function (foundClass) {
+      if (!foundClass) {
+        res.status(404).json({message: "Class not found"});
+      }
+      else {
+        Student.destroy({
+            where: {
+              username : req.params.studentId,
+              ClassId : req.params.classId
+            }
+          })
+          .then(function (result) {
+            if (!result) {
+              res.status(404).json({message: "Student not found"});
+            }
+            else {
+              res.json({message: "Student deleted"});
+            }
+          });
+      }
+    });
   }
   catch (e) {
     console.error(e);
